% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Match&Permutation_Functions.R
\name{angleMatchICA}
\alias{angleMatchICA}
\title{Match the colums of Mx and My}
\usage{
angleMatchICA(Mx, My, Sx = NULL, Sy = NULL)
}
\arguments{
\item{Mx}{Subject score for X   matrix of n x n.comp}

\item{My}{Subject score for Y   matrix of n x n.comp}

\item{Sx}{Variable loadings for X  matrix of n.comp x px}

\item{Sy}{Variable loadings for Y  matrix of n.comp x py}
}
\value{
a list of matrixes:
 ## Mx:
 ## My:
 ## matchedangles:
 ## allangles:
 ## perm:
 ## omangles:
}
\description{
\code{angleMatchICA} match the colums of Mx and My, using the n x p parameterization of the JIN decomposition assumes
}
