% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_functions.r
\name{aveM}
\alias{aveM}
\title{Average Mj for Mx and My
Here subjects are by rows, columns correspond to components}
\usage{
aveM(mjX, mjY)
}
\arguments{
\item{mjX}{n x rj}

\item{mjY}{n x rj}
}
\value{
a new Mj
}
\description{
Average Mj for Mx and My
Here subjects are by rows, columns correspond to components
}
\examples{
\donttest{
#get simulation data
data(exampledata)
data=exampledata

# To get n.comp value, we can use NG_number function.

# use JB statistic as the measure of nongaussianity to run lngca with df=0
output_JB=singR(dX=exampledata$dX,dY=exampledata$dY,
df=0,rho_extent="small",distribution="JB",individual=TRUE)

est.Mj = aveM(outputJB$est.Mjx,outputJB$est.Mjy)

}
}
