% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_functions.r
\name{calculateJB}
\alias{calculateJB}
\title{JB score calculate}
\usage{
calculateJB(S = NULL, U = NULL, X = NULL, alpha = 0.8)
}
\arguments{
\item{S}{the variable loadings r x px.}

\item{U}{U matrix for matched columns rj x n}

\item{X}{whitened data matrix n x px, data = whitenerXA \%*\% dXcentered}

\item{alpha}{default = 0.8}
}
\value{
the sum of JB score for each component, the row of Sj.
}
\description{
JB score calculate
}
