% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdnet_tools.R
\name{cmd_birdnet}
\alias{cmd_birdnet}
\title{Generate command to run BirdNET Analyzer}
\usage{
cmd_birdnet(mode = "analyze", birdnet_loc, birdnet_args)
}
\arguments{
\item{mode}{The mode in which BirdNET Analyzer should be run (such as "analyze" or "embeddings").}

\item{birdnet_loc}{The location of the BirdNET Analyzer.}

\item{birdnet_args}{A list of command line arguments to be passed to BirdNET Analyzer.}
}
\value{
A list containing the generated command file location and the output file path.
}
\description{
This function generates a command to run BirdNET Analyzer with the specified mode,
BirdNET location, and command line arguments.
}
\examples{

# Set the location of BirdNET Analyzer
birdnet_location <- "/home/user/BirdNET-Analyzer/"

# Set command line arguments
birdnet_arguments <- list(
  i = "path/to/input/file",
  o = "path/to/output/folder",
  lat = 40.7128,
  lon = -74.0060,
  week = 23,
  slist = "path/to/species/list",
  sensitivity = 1.0,
  min_conf = 0.1,
  overlap = 0.0,
  rtype = "table",
  threads = 4,
  batchsize = 1,
  locale = "en",
  sf_thresh = 0.03,
  classifier = "path/to/custom/classifier",
  fmin = 0,
  fmax = 15000
)

# Generate BirdNET command
birdnet_command <- cmd_birdnet(mode = "analyze", birdnet_loc = birdnet_location, birdnet_args = birdnet_arguments)

# View the generated command and output file path
print(birdnet_command)

}
