% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jngcaFunctions.R
\name{covwhitener}
\alias{covwhitener}
\title{Returns square root of the precision matrix for whitening}
\usage{
covwhitener(X, n.comp = ncol(X), center.row = FALSE)
}
\arguments{
\item{X}{Matrix}

\item{n.comp}{the number of components}

\item{center.row}{whether to center}
}
\value{
square root of the precision matrix for whitening
}
\description{
Returns square root of the precision matrix for whitening
}
