% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_path_info.R
\name{create_id_from_path}
\alias{create_id_from_path}
\title{Generate an ID from a given path}
\usage{
create_id_from_path(path, ...)
}
\arguments{
\item{path}{The path from which to generate the ID.}

\item{...}{Additional arguments to be passed to extract_path_info().}
}
\value{
An ID generated from the path.
}
\description{
This function generates an ID based on the provided path and specified levels.
}
\examples{

# Provide a path
path <- "path/to/your/file"

# Generate an ID from the path
id <- create_id_from_path(path)

# View the generated ID
print(id)

}
