% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateFunctions.R
\name{curvilinear_c}
\alias{curvilinear_c}
\title{Curvilinear algorithm based on C code with r0 joint components}
\usage{
curvilinear_c(
  Ux,
  Uy,
  xData,
  yData,
  invLx,
  invLy,
  rho,
  tau = 0.01,
  alpha = 0.8,
  maxiter = 1000,
  tol = 1e-06,
  r0
)
}
\arguments{
\item{Ux}{Matrix with n.comp x n, initial value of Ux, comes from greedyMatch.}

\item{Uy}{Matrix with n.comp x n, initial value of Uy, comes from greedyMatch.}

\item{xData}{matrix with n x px, Xw = Lx \%*\% Xc.}

\item{yData}{matrix with n x py, Yw = Ly \%*\% Yc.}

\item{invLx}{Inverse matrix of Lx, matrix n x n.}

\item{invLy}{Inverse matrix of Ly, matrix n x n.}

\item{rho}{the weight parameter of matching relative to non-gaussianity.}

\item{tau}{initial step size, default value is 0.01}

\item{alpha}{default value is 0.8}

\item{maxiter}{default value is 1000}

\item{tol}{the threshold of change in Ux and Uy to stop the curvlinear function}

\item{r0}{the joint rank, comes from greedyMatch.}
}
\value{

}
\description{
Curvilinear algorithm based on C code with r0 joint components
}
