% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tools.R
\name{db_init}
\alias{db_init}
\title{function to initialize database and users based on sql script}
\usage{
db_init(
  host = "localhost",
  database = "postgres",
  user = "postgres",
  port = 5432
)
}
\arguments{
\item{host}{character, host address, default is "localhost"}

\item{database}{character, database name for initial connection, default is "postgres"}

\item{user, }{character, name of the user with rights to create user roles and database, default is "postgres"}

\item{port, }{numeric, port for database connection}
}
\description{
function to initialize database and users based on sql script
}
