% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tools.R
\name{db_structure}
\alias{db_structure}
\title{function to initialize schemas and tables after ctm_db has been setup by create_ctm_}
\usage{
db_structure(host = "localhost", user = "postgres", port = 5432)
}
\arguments{
\item{host}{character, host address, default is "localhost"}

\item{user}{character, name of the user with rights to create tables and schemas, default is "postgres"}

\item{port}{integer, port for database connection}
}
\description{
function to initialize schemas and tables after ctm_db has been setup by create_ctm_
}
