% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tools.R
\name{db_upsert}
\alias{db_upsert}
\title{Generic function to update information in database}
\usage{
db_upsert(upload_data, con, schema, table, idfields, verbose = TRUE)
}
\arguments{
\item{upload_data}{'data.frame', values to be updated, does not need to
include all columns in the database}

\item{con}{postgresqlconnection}

\item{schema}{'character', the schema name to receive the updated
values}

\item{table}{'character', the table name to receive the updated
values}

\item{idfields}{'character', one or more id fields that are present
in both the 'upload_data' and the database table, these cannot change}

\item{verbose}{'logical', be verbose about operation, default true}
}
\value{
logical, whether 'nrow(upload_data)' rows were affected; if anfa
  error occurred, it is messaged to the console and a `FALSE` is
  returned
}
\description{
Generic function to update information in database
}
