% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_tools.R
\name{download_categories}
\alias{download_categories}
\title{function to download categories}
\usage{
download_categories(conn)
}
\arguments{
\item{conn}{database connection}
}
\description{
function to download categories
}
