% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lngca.R
\name{est.M.ols}
\alias{est.M.ols}
\title{Estimate mixing matrix from estimates of components}
\usage{
est.M.ols(sData, xData, intercept = TRUE)
}
\arguments{
\item{sData}{S rx x px}

\item{xData}{dX n x px}

\item{intercept}{default = TRUE}
}
\value{
a matrix Mx, dimension n x rx.
}
\description{
Estimate mixing matrix from estimates of components
}
