% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{exampledata}
\alias{exampledata}
\title{Data for simulation example 1}
\format{
A data list with 10 subsets:
\describe{
     \item{dX}{original data matrix for X with n x px, 48x3602}
     \item{dY}{original data matrix for Y with n x py, 48x4950}
     \item{mj}{true mj matrix, n x rj, 48x2}
     \item{sIx}{true S matrix of independent non-Gaussian components in X, ri_x x px, 2x3602}
     \item{sIy}{true S matrix of independent non-Gaussian components in Y, ri_y x py, 2x4950}
     \item{sjx}{true S matrix of joint non-Gaussian components in X, rj x px, 2x3602}
     \item{sjy}{true S matrix of joint non-Gaussian components in Y, rj x py, 2x4950}
     \item{snr}{signal to noise ratio}
     \item{R2x}{R2 for x data}
     \item{R2y}{R2 for y data}
     }
}
\usage{
exampledata
}
\description{
Data for simulation example 1
}
\keyword{datasets}
