% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_path_info.R
\name{extract_path_info}
\alias{extract_path_info}
\title{Extract path information}
\usage{
extract_path_info(path, level = "data")
}
\arguments{
\item{path}{The path from which to extract information.}

\item{level}{The level of information to extract ("data", "data_dir", "deployment_id", "location_id", "region_id", "project_id").}
}
\value{
A data frame containing extracted information.
}
\description{
This function extracts information from a given path up to a specified level.
}
\examples{

# Provide a path
path <- "path/to/your/file"

# Extract information from the path up to the "data" level
extracted_info <- extract_path_info(path, level = "data")

# View the extracted information
print(extracted_info)

}
