% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FOBIasymp.R
\name{FOBIasymp.2}
\alias{FOBIasymp.2}
\title{Use FOBIasymp to find non-Gaussian components number}
\usage{
FOBIasymp.2(
  X,
  k,
  type = "S3",
  model = "NGCA",
  method = "satterthwaite",
  whiten = c("sqrtprec", "eigenvec")
)
}
\arguments{
\item{X}{input data with p x n.}

\item{k}{the non-Gaussian components number in Null hypothesis.}

\item{type}{'S1', 'S2' or 'S3'.}

\item{model}{'NGCA' or 'ICA'}

\item{method}{"satterthwaite","integration","saddlepoint"}

\item{whiten}{'sqrtprec','eigenvec'}
}
\value{
p-value for the test
}
\description{
Use FOBIasymp to find non-Gaussian components number
}
\examples{
data(exampledata)
FOBIasymp.2(t(exampledata$dX),k=4)
}
