% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployment_info.R
\name{format_deployment}
\alias{format_deployment}
\title{Format deployment data from a file}
\usage{
format_deployment(file, device = "songmeter")
}
\arguments{
\item{file}{The path to the deployment summary file.}

\item{device}{The type of device used for deployment (default is "songmeter").}
}
\value{
A data frame containing formatted deployment information.
}
\description{
This function reads and formats deployment data from a file, specifically designed for Songmeter devices.
}
