% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jngcaFunctions.R
\name{frobICA}
\alias{frobICA}
\title{Match mixing matrices}
\usage{
frobICA(M1 = NULL, M2 = NULL, S1 = NULL, S2 = NULL, standardize = FALSE)
}
\arguments{
\item{M1}{Subject score 1}

\item{M2}{Subject score 2}

\item{S1}{Loading 1}

\item{S2}{Loading 2}

\item{standardize}{whether to standardize}
}
\value{

}
\description{
Match mixing matrices
}
