% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustic_indices.R
\name{get_files_info}
\alias{get_files_info}
\title{get files info for running acoustic indices analysis}
\usage{
get_files_info(con)
}
\description{
get files info for running acoustic indices analysis
}
