% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/give_it_up.R
\name{give_up}
\alias{give_up}
\title{Generate string saying "give something up"}
\usage{
give_up(what, who = "")
}
\arguments{
\item{what}{string specifying what should be given up}

\item{who}{string specifying who should give up something}
}
\value{
string
}
\description{
Generate string saying "give something up"
}
\examples{
give_up('it')
give_up('it', 'Baby')
}

