% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_functions.R, R/Basic_functions.r
\name{\%^\%}
\alias{\%^\%}
\title{Calculate the power of a square matrix}
\usage{
S \%^\% power

S \%^\% power
}
\arguments{
\item{S}{a square matrix}

\item{power}{the times of power}
}
\value{
a matrix after power calculation that eigenvector x diag(eigenvalue ^ power) x eigenvector'

a matrix after power calculation that eigenvector x diag(eigenvalue ^ power) x eigenvector'
}
\description{
returns a matrix composed of eigenvector x diag(eigenvalue ^ power) x eigenvector'

returns a matrix composed of eigenvector x diag(eigenvalue ^ power) x eigenvector'
}
\examples{
a <- matrix(1:9,3,3)
a \%^\% 2

a <- matrix(1:9,3,3)
a \%^\% 2

}
