% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mCCAjointICA.R
\name{jointICA}
\alias{jointICA}
\title{Find the joint component through joint_ICA method.}
\usage{
jointICA(dX, dY, r0 = 2)
}
\arguments{
\item{dX}{the first dataset.}

\item{dY}{the second dataset.}

\item{r0}{the number of joint components.}
}
\value{
a list with matrices
## S: the concatenated matrix of SjX and SjY, [SjX',SjY'],at here (pX+pY) x rj
## Mjoint: subject score estimated by joint ICA, at here rj x nsubject. for the example dataset, the rj is 2.
}
\description{
Find the joint component through joint_ICA method.
}
\examples{
out_jointICA <- jointICA(dXcentered, dYcentered, r0 = 2).
}
