% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jngcaFunctions.R
\name{lngca}
\alias{lngca}
\title{Decomposite the original data through LNGCA method}
\usage{
lngca(
  xData,
  n.comp = ncol(xData),
  W.list = NULL,
  whiten = c("eigenvec", "sqrtprec", "none"),
  maxit = 1000,
  eps = 1e-06,
  verbose = FALSE,
  restarts.pbyd = 0,
  restarts.dbyd = 0,
  distribution = c("tiltedgaussian", "logistic", "JB"),
  density = FALSE,
  out.all = FALSE,
  orth.method = c("svd", "givens"),
  reinit.max.comp = FALSE,
  max.comp = FALSE,
  df = 0,
  irlba = FALSE,
  ...
)
}
\arguments{
\item{xData}{the original dataset for decomposition, matrix of px x n}

\item{n.comp}{the number of components need to be estimated.}

\item{whiten}{whiten method with default}

\item{maxit}{defalut = 1000}

\item{eps}{default = 1e-06}

\item{verbose}{default = FALSE}

\item{restarts.pbyd}{default = 0}

\item{restarts.dbyd}{default = 0}

\item{distribution}{distribution methods with default}

\item{density}{default = FALSE}

\item{out.all}{default = FALSE}

\item{orth.method}{default = 'svd'}

\item{max.comp}{logical variable that estiamtes the max number of non-gaussian components.}

\item{df}{default = 0}

\item{irlba}{whether require irlba package.}

\item{...}{}
}
\value{
Function outputs a list including the following:
\describe{
      \item{\code{Ws}}{ t(Ux) matrix n x n.comp, part of the expression that Ax = Ux x Lx and Ax x Xc = Sx, where Lx is the whitener matrix.}
      \item{\code{loglik}}{the value of log-likelihood in the lngca method.}
      \item{\code{S}}{the variable loading matrix px x n.comp, each column is a component, which can be used to measure nongaussianity}
      \item{\code{df}}{egree of freedom.}
      \item{\code{distribution}}{the method used for data decomposition.}
      \item{\code{whitener}}{A symmetric whitening matrix n x n from dX, the same with  whitenerXA = est.sigmaXA \%^\% -0.5}
      \item{\code{M}}{Mtrix with n.comp x n.}
      \item{\code{nongaussianity}}{the nongaussianity score for each component saved in S matrix.}
}
}
\description{
Decomposite the original data through LNGCA method
}
\examples{

}
