% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jngcaFunctions.R
\name{lngca_multicore}
\alias{lngca_multicore}
\title{Title}
\usage{
lngca_multicore(
  xData,
  n.comp = ncol(xData),
  ncores = 20,
  W.list = NULL,
  whiten = c("eigenvec", "none"),
  restarts = NULL,
  distribution = c("tiltedgaussian", "logistic", "JB"),
  df = 0,
  initFOBI = FALSE,
  scratch = "~/temp",
  keepall = FALSE,
  sourcefile = "~/Dropbox/JINGCA/Programs/Functions/jngcaFunctions.R",
  ...
)
}
\arguments{
\item{xData}{input data}

\item{n.comp}{number of data columns}

\item{ncores}{number of cores, default = 20}

\item{W.list}{list}

\item{whiten}{whiten method 'eigenvec' or 'none'}

\item{restarts}{default = ncores}

\item{distribution}{type of distribution}

\item{df}{dataframe}

\item{initFOBI}{default = False}

\item{scratch}{temporary directory name}

\item{sourcefile}{the filepath contains jngca_functions}

\item{...}{}
}
\value{
a list with matrices
## Ws:
## lgolik:
## S: Variable loadings
## df:
## distribution: the distribution method used in mlcaFP
## whitener:
## M:
## nongaussianity: the value of nan-gaussianity for each component
}
\description{
Title
}
