% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Match&Permutation_Functions.R
\name{matchICA}
\alias{matchICA}
\title{match ICA}
\usage{
matchICA(S, template, M = NULL)
}
\arguments{
\item{S}{loading variable matrix}

\item{template}{template for match}

\item{M}{subject score matrix}
}
\value{
the match result
}
\description{
match ICA
}
