% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jngcaFunctions.R
\name{mProDenICA}
\alias{mProDenICA}
\title{multiple ProdenICA}
\usage{
mProDenICA(
  X,
  n.comp = ncol(X),
  restarts = 0,
  tol = 1e-07,
  maxit = 100,
  G = c("GPois", "G0", "G1"),
  verbose = FALSE,
  whiten = FALSE,
  ...
)
}
\arguments{
\item{X}{Original data}

\item{n.comp}{the max components}

\item{restarts}{the number of restart}

\item{tol}{tol amount}

\item{maxit}{the max iteration number}

\item{G}{G method}

\item{verbose}{default = false}

\item{whiten}{whether to whiten}

\item{...}{ellipsis}
}
\value{

}
\description{
multiple ProdenICA
}
