% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{newSimFMRI}
\alias{newSimFMRI}
\title{New dataset for simulation of FMRI}
\usage{
newSimFMRI(
  snr = 1,
  noisyICA = FALSE,
  nTR = 50,
  nImages = 1,
  phi = 0.5,
  dim.data = c(33, 33),
  var.inactive = 1e-04
)
}
\arguments{
\item{snr}{numeric, which shows the ratio of non-Gaussian (both joint and individual) to Gaussian components in dataset.}

\item{noisyICA}{whether to use noisyICA method to simulate gaussian noise.}

\item{nTR}{the number of time points, which is not changeable at here.}

\item{nImages}{the number of images for each component generated in the dataset.}

\item{phi}{default value is 0.5.}

\item{dim.data}{vector of length 2, the dimension of the simulation data.}

\item{var.inactive}{the background variance of non-active voxels or edges within a non-Gaussian component. Setting this equal to zero results in true sparsity in a non-Gaussian components.}
}
\description{
\code{newSimFMRI} Create a dataset containing 4 components correspond to 2D images, roughly similar to a brain activation map.
}
