% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic functions.r
\name{orthogonalize}
\alias{orthogonalize}
\title{Orthogonalization of matrix}
\usage{
orthogonalize(W)
}
\arguments{
\item{W}{arbitrary matrix}
}
\value{
orthogonalized matrix
}
\description{
Orthogonalization of matrix
}
