% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Match&Permutation_Functions.R
\name{permTestJointRank}
\alias{permTestJointRank}
\title{Permutation test with Greedymatch}
\usage{
permTestJointRank(
  MatchedMx,
  MatchedMy,
  nperm = 1000,
  alpha = 0.01,
  multicore = 0
)
}
\arguments{
\item{MatchedMx}{matrix with nsubject x n.comp.X, comes from greedymatch}

\item{MatchedMy}{matrix with nsubject2 x n.comp.Y, comes from greedymatch}

\item{nperm}{default value = 1000}

\item{alpha}{default value = 0.01}

\item{multicore}{default value = 0}
}
\value{
a list of matrixes
 ## rj: joint component rank
 ## pvalues: pvalue for the components(columns) not matched
 ## fwer_alpha: quantile of corr permutation with 1- alpha
}
\description{
Permutation test with Greedymatch
}
