% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jngcaFunctions.R
\name{plotNetwork}
\alias{plotNetwork}
\title{Title}
\usage{
plotNetwork(
  component,
  title = "",
  qmin = 0.005,
  qmax = 0.995,
  path = "~/Dropbox/JINGCA/Data/community_affiliation_mmpplus.csv",
  make.diag = NA
)
}
\arguments{
\item{component}{component for the plot, px x rx}

\item{title}{title for plot}

\item{qmin}{default value = 0.005}

\item{qmax}{default value = 0.995}

\item{path}{path for the saved plot data}

\item{make.diag}{default value = NA}
}
\value{

}
\description{
Title
}
