% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/All_functions.R
\name{plotNetwork_change}
\alias{plotNetwork_change}
\title{Plot for the changed example of Y component.}
\usage{
plotNetwork_change(
  component,
  title = "",
  qmin = 0.005,
  qmax = 0.995,
  path = "~/Dropbox/JINGCA/Data/community_affiliation_mmpplus.csv",
  make.diag = NA
)
}
\arguments{
\item{component}{component for the plot, px x rx}

\item{title}{title for plot}

\item{qmin}{default value = 0.005}

\item{qmax}{default value = 0.995}

\item{path}{default path}

\item{make.diag}{default value = NA}
}
\value{
a list as followed:
\describe{
      \item{\code{netmatfig}}{component loadings from rs correlation}
      \item{\code{loadingsfig}}{the sum of the absolute values of the rows of netmatfig}
      \item{\code{netmat}}{a matrix can be used with image() function}
      \item{\code{loadingsummary}}{loadings summary for each row of netmatfig}
}
}
\description{
Plot for the changed example of Y component.
}
