% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jngcaFunctions.R
\name{pmse}
\alias{pmse}
\title{Match mixing matrices}
\usage{
pmse(M1 = NULL, M2 = NULL, S1 = NULL, S2 = NULL, standardize = FALSE)
}
\arguments{
\item{M1}{Subject score 1 matrix r x n.}

\item{M2}{Subject score 2 matrix r x n.}

\item{S1}{Loading 1 with matrix p x r.}

\item{S2}{Loading 2 with matrix p x r.}

\item{standardize}{whether to standardize}
}
\description{
Match mixing matrices
}
