% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acoustic_indices.R
\name{proc_acoustic}
\alias{proc_acoustic}
\title{function to calculate multiple acoustic indices}
\usage{
proc_acoustic(files_info)
}
\arguments{
\item{files_info}{vector created from exif data}
}
\description{
function to calculate multiple acoustic indices
}
