% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{simdata}
\alias{simdata}
\title{Data for simulation example}
\format{
A data list with 7 subsets:
\describe{
     \item{dX}{original data matrix for X with n x px, 48x3602}
     \item{dY}{original data matrix for Y with n x py, 48x4950}
     \item{mj}{true mj matrix, n x rj, 48x2}
     \item{sIx}{true S matrix of independent non-Gaussian components in X, ri_x x px, 2x3602}
     \item{sIy}{true S matrix of independent non-Gaussian components in Y, ri_y x py, 2x4950}
     \item{sjx}{true S matrix of joint non-Gaussian components in X, rj x px, 2x3602}
     \item{sjy}{true S matrix of joint non-Gaussian components in Y, rj x py, 2x4950}

     }
}
\usage{
simdata
}
\description{
Data for simulation example
}
\keyword{datasets}
