% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{SimFMRI123}
\alias{SimFMRI123}
\title{Sim data generation for 123 graph}
\usage{
SimFMRI123(
  snr = 1,
  noisyICA = FALSE,
  nTR = 50,
  nImages = 1,
  phi = 0.5,
  dim.data = c(33, 33),
  var.inactive = 1e-04
)
}
\arguments{
\item{snr}{the signal to noise ratio}

\item{noisyICA}{whether use noisyICA}

\item{nTR}{default value is 50}

\item{nImages}{the number of images, default value is 1}

\item{phi}{default value = 0.5}

\item{dim.data}{dimension of data matrix}

\item{var.inactive}{default value is 0.0001}
}
\value{
a list of simulation data
}
\description{
Sim data generation for 123 graph
}
