% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sing.R
\name{sing}
\alias{sing}
\title{Writing a number of syllabus on standard output defined by \code{speed} number of seconds}
\usage{
sing(text, speed = 4)
}
\arguments{
\item{text}{character vector}

\item{speed}{numeric}
}
\value{
Nothing is returned, the text is written to \code{stdout}.
}
\description{
Writing a number of syllabus on standard output defined by \code{speed} number of seconds
}

