% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singR.R
\name{singR}
\alias{singR}
\title{sing method for data integration}
\usage{
singR(
  dX,
  dY,
  n.comp = 12,
  df = 0,
  rho_extent = c("small", "medium", "large"),
  Cplus = T,
  tol = 1e-10,
  stand = F,
  distribution = "JB"
)
}
\arguments{
\item{dX}{original dataset for decomposition, matrix of n x px.}

\item{dY}{original dataset for decomposition, matrix of n x py.}

\item{n.comp}{the component number used in separate lngca process.}

\item{df}{default value=0 when use JB, df>0 when use tileted gaussian.}

\item{rho_extent}{small,medium or large.}

\item{Cplus}{whether to use C code in curvilinear search.}

\item{tol}{difference tolerance in curvilinear search.}

\item{stand}{whether to use standardization, if it was true, it will make the column and row means to 0 and columns sd to 1. If false, it will only make the row means to 0.}

\item{distribution}{"JB" or "tiltedgaussian"}
}
\value{
Function outputs a list including the following:
\describe{
      \item{\code{Sx}}{variable loadings for dataset X with matrix r x px.}
      \item{\code{Sy}}{variable loadings for dataset Y with matrix r x py.}
      \item{\code{Mxjoint}}{Mj of X data.}
      \item{\code{Myjoint}}{Mj of Y data.}
      \item{\code{est.Mj}}{estimated Mj for both datasets.}
      \item{\code{C_plus}}{whether to use C version of curvilinear search}
      \item{\code{rho_extent}}{the weight of rho in search}
      \item{\code{df}}{degree of freedom, = 0 when use JB,>0 when use tiltedgaussian}
}
}
\description{
sing method for data integration
}
\examples{
#get simulation data
data(exampledata)
data=exampledata

# use JB stat to compute with singR
output_JB=singR(dX=data$dX,dY=data$dY,n.comp=12,df=0,rho_extent="small",distribution="JB")
output_tilted=singR(dX=data$dX,dY=data$dY,n.comp=12,df=0,rho_extent="small",distribution="tiltedgaussian")

# use frobICA to show the difference of JB and tiltedgaussian
frobICA_JB=frobICA(M1 = t(output_JB$est.Mj),M2 = t(data$mj),standardize = T) #0.0071682
frobICA_tilted=frobICA(M1 = t(output_tilted$est.Mj),M2 = t(data$mj),standardize = T) #0.0071295

}
