% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/song_info.R
\name{song_info}
\alias{song_info}
\title{Extract metadata information from audio files in a directory}
\usage{
song_info(
  song_dir = here::here(),
  info_out = tempfile("out", tmpdir = here::here(), fileext = ".csv")
)
}
\arguments{
\item{song_dir}{The directory containing audio files.}

\item{info_out}{The output file path to save the metadata information (default is a temporary CSV file).}
}
\value{
A data frame containing metadata information extracted from audio files.
}
\description{
This function uses ExifTool to extract metadata information from audio files in a directory
and saves the information in a CSV file.
}
\examples{

# Provide the path to the directory containing audio files
song_directory <- "path/to/your/audio/files"

# Extract metadata information
song_info <- song_info(song_directory)

# View the extracted metadata information
print(song_info)

}
