% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tap_along.R
\name{tap.along}
\alias{tap.along}
\title{Type lyrics and echo a beep sound along songs defined by four parameters}
\usage{
tap.along(time.pattn, tempo = 60, lyrics = NA, repet = 4)
}
\arguments{
\item{time.pattn}{character vector}

\item{tempo}{numeric}

\item{lyrics}{character vector}
}
\value{
Nothing is returned, the text is written to \code{stdout}.
}
\description{
Type lyrics and echo a beep sound along songs defined by four parameters
}
\examples{
heyjude <- list()
heyjude$lyrics <- c(rep(NA, 11), "Hey", "Hey","Jude")
heyjude$time.pattn <- "h, q, q, s, s, e, hd, s, s, e, h, s, ed, w"
heyjude$tempo <- 60
heyjude$repet <- 2 # I hesitated between this and "Inf"
do.call(what = tap.along, args = heyjude)
}

