% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_functions.r
\name{theta2W}
\alias{theta2W}
\title{Convert angle vector into  orthodox matrix}
\usage{
theta2W(theta)
}
\arguments{
\item{theta}{vector of angles theta}
}
\value{

}
\description{
Convert angle vector into  orthodox matrix
}
