% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_functions.r
\name{tiltedgaussian}
\alias{tiltedgaussian}
\title{tiltedgaussian}
\usage{
tiltedgaussian(xData, df = 8, B = 100, ...)
}
\arguments{
\item{xData}{input data}

\item{df}{degree freedom}

\item{B}{default value=100}

\item{...}{}
}
\description{
tiltedgaussian
}
