% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jngcaFunctions.R
\name{vec2net}
\alias{vec2net}
\title{Create network matrices from vectorized lower diagonals
\code{vec2net} transfer the matrix vectorized lower diagonals into net to show the component image.}
\usage{
vec2net(invector, make.diag = 1)
}
\arguments{
\item{invector}{vectorized lower diagonals.}

\item{make.diag}{default value = 1.}
}
\value{

}
\description{
Create network matrices from vectorized lower diagonals
\code{vec2net} transfer the matrix vectorized lower diagonals into net to show the component image.
}
