% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/give_it_up.R
\name{verify}
\alias{verify}
\title{Generate string saying if "someone can do something"}
\usage{
verify(what, who = "", times = 1)
}
\arguments{
\item{what}{string specifying what can be done}

\item{who}{string specifying who can do the thing}

\item{times}{numeric value representing the number of repetitions}
}
\value{
string
}
\description{
Generate string saying if "someone can do something"
}
\examples{
verify(give_up(what = 'it', who = 'you'))
verify(give_up(what = 'it'), who = 'you', times = 2)
}

