% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic functions.r
\name{whitener}
\alias{whitener}
\title{Whitening Function}
\usage{
whitener(X, n.comp = ncol(X), center.row = FALSE, irlba = FALSE)
}
\arguments{
\item{X}{dataset}

\item{irlba}{}
}
\value{

}
\description{
Whitening Function
}
