% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_functions.r
\name{calculateJB}
\alias{calculateJB}
\title{Calculates the sum of the JB scores across all components, useful for determining rho.
The data has to be standardized and mean 0 and sd to 1.}
\usage{
calculateJB(S = NULL, U = NULL, X = NULL, alpha = 0.8)
}
\arguments{
\item{S}{the variable loadings r x px.}

\item{U}{U matrix for matched columns rj x n}

\item{X}{whitened data matrix n x px, data = whitenerXA \%*\% dXcentered}

\item{alpha}{JB weighting of skewness and kurtosis. default = 0.8}
}
\value{
the sum of JB score across all components.
}
\description{
Calculates the sum of the JB scores across all components, useful for determining rho.
The data has to be standardized and mean 0 and sd to 1.
}
