% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jngcaFunctions.R
\name{create.graph.long}
\alias{create.graph.long}
\title{create graph with netmat and mmp_order}
\usage{
create.graph.long(gmatrix, sort_indices = NULL)
}
\arguments{
\item{gmatrix}{netmat}

\item{sort_indices}{mmp_order}
}
\value{
a data.frame called meltsub
}
\description{
create graph with netmat and mmp_order
}
