% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployment_info.R
\name{deployment_info}
\alias{deployment_info}
\title{get deployment info}
\usage{
deployment_info(input_dir)
}
\description{
specific function for songmeter micro txt files which looks for files for each deployment named [RECORDER_NAME]_Summary.txt
}
