% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jngcaFunctions.R
\name{est.M.ols}
\alias{est.M.ols}
\title{Title}
\usage{
est.M.ols(sData, xData, intercept = TRUE)
}
\arguments{
\item{sData}{t(S) px x n.comp}

\item{xData}{t(dX) px x n}

\item{intercept}{default = TRUE}
}
\value{
a matrix t(Mx) n.comp x n
}
\description{
Title
}
