% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_functions.r
\name{gen.inits}
\alias{gen.inits}
\title{Generate initialization from specific space}
\usage{
gen.inits(p, d, runs, orth.method = c("svd", "givens"))
}
\arguments{
\item{p}{p*p orthodox matrix}

\item{d}{p*d orthodox matrix}

\item{runs}{the number of orthodox matrix}

\item{orth.method}{orthodox method}
}
\value{
a list of initialization of mixing matrices.
}
\description{
Generate initialization from specific space
}
\examples{
gen.inits(2,3,3,'svd')
}
