% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jngcaFunctions.R
\name{lngca}
\alias{lngca}
\title{Decomposite the original data through LNGCA method}
\usage{
lngca(
  xData,
  n.comp = ncol(xData),
  W.list = NULL,
  whiten = c("eigenvec", "sqrtprec", "none"),
  maxit = 1000,
  eps = 1e-06,
  verbose = FALSE,
  restarts.pbyd = 0,
  restarts.dbyd = 0,
  distribution = c("tiltedgaussian", "logistic", "JB"),
  density = FALSE,
  out.all = FALSE,
  orth.method = c("svd", "givens"),
  reinit.max.comp = FALSE,
  max.comp = FALSE,
  df = 0,
  irlba = FALSE,
  ...
)
}
\arguments{
\item{xData}{the original dataset for decomposition, matrix of px x n}

\item{n.comp}{the number of components need to be estimated.}

\item{whiten}{whiten method with default}

\item{maxit}{defalut = 1000}

\item{eps}{default = 1e-06}

\item{verbose}{default = FALSE}

\item{restarts.pbyd}{default = 0}

\item{restarts.dbyd}{default = 0}

\item{distribution}{distribution methods with default}

\item{density}{default = FALSE}

\item{out.all}{default = FALSE}

\item{orth.method}{default = 'svd'}

\item{max.comp}{logical variable that estiamtes the max number of non-gaussian components.}

\item{df}{default = 0}

\item{irlba}{whether require irlba package.}

\item{...}{}
}
\value{

}
\description{
Decomposite the original data through LNGCA method
}
\examples{

}
