% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lngca.R
\name{lngca}
\alias{lngca}
\title{Decomposite the original data through LNGCA method}
\usage{
lngca(
  xData,
  n.comp = ncol(xData),
  W.list = NULL,
  whiten = c("eigenvec", "sqrtprec", "none"),
  maxit = 1000,
  eps = 1e-06,
  verbose = FALSE,
  restarts.pbyd = 0,
  restarts.dbyd = 0,
  distribution = c("tiltedgaussian", "logistic", "JB"),
  density = FALSE,
  out.all = FALSE,
  orth.method = c("svd", "givens"),
  reinit.max.comp = FALSE,
  max.comp = FALSE,
  df = 0,
  stand = FALSE,
  ...
)
}
\arguments{
\item{xData}{the original dataset for decomposition, matrix of px x n}

\item{n.comp}{the number of components need to be estimated.}

\item{W.list}{list}

\item{whiten}{whiten method with default}

\item{maxit}{max iteration, defalut = 1000}

\item{eps}{default = 1e-06}

\item{verbose}{default = FALSE}

\item{restarts.pbyd}{default = 0}

\item{restarts.dbyd}{default = 0}

\item{distribution}{distribution methods with default}

\item{density}{default = FALSE}

\item{out.all}{default = FALSE}

\item{orth.method}{default = 'svd'}

\item{reinit.max.comp}{default = FALSE}

\item{max.comp}{logical variable that estiamtes the max number of non-gaussian components.}

\item{df}{default = 0}

\item{stand}{whether to standard the data}

\item{...}{ellipsis}
}
\value{
Function outputs a list including the following:
\describe{
      \item{\code{U}}{ matrix rx x n, part of the expression that Ax = Ux x Lx and Ax x Xc = Sx, where Lx is the whitener matrix.}
      \item{\code{loglik}}{the value of log-likelihood in the lngca method.}
      \item{\code{S}}{the variable loading matrix r x px, each row is a component, which can be used to measure nongaussianity}
      \item{\code{df}}{egree of freedom.}
      \item{\code{distribution}}{the method used for data decomposition.}
      \item{\code{whitener}}{A symmetric whitening matrix n x n from dX, the same with  whitenerXA = est.sigmaXA \%^\% -0.5}
      \item{\code{M}}{Mx Mtrix with n x rx.}
      \item{\code{nongaussianity}}{the nongaussianity score for each component saved in S matrix.}
}
}
\description{
Decomposite the original data through LNGCA method
}
\examples{
\dontrun{
#get simulation data
data(exampledata)
data=exampledata

# use JB statistic as the measure of nongaussianity to run lngca with df=0
estX_JB = lngca(xData = t(data$dX), n.comp = 4, whiten = 'sqrtprec', restarts.pbyd = 20, distribution='JB',df=0)

# use the tiltedgaussian distribution to run lngca with df=3
estX_tilt = lngca(xData = t(data$dX), n.comp = 4, whiten = 'sqrtprec', restarts.pbyd = 20, distribution='tiltedgaussian',df=3)

# true non-gaussian component of Sx, includ individual and joint components
trueSx = rbind(data$sjX,data$siX)

# use frobICA to compare the difference of the two methods
frobICA(S1 = t(trueSx),S2=t(estX_JB$S),standardize = TRUE) #0.2341096
frobICA(S1 = t(trueSx),S2=t(estX_tilt$S),standardize = TRUE) #0.1824922

# the lngca using tiltedgaussian is more accurate with smaller frobICA value. But it will cost more time.
}

}
