% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{newSimFMRI}
\alias{newSimFMRI}
\title{New dataset for simulation of FMRI}
\usage{
newSimFMRI(
  snr = 1,
  noisyICA = FALSE,
  nTR = 50,
  nImages = 1,
  phi = 0.5,
  dim.data = c(33, 33),
  var.inactive = 1e-04
)
}
\description{
New dataset for simulation of FMRI
}
