% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Match&Permutation_Functions.R
\name{NG_number}
\alias{NG_number}
\title{find the number of non-Gaussian components in the data.}
\usage{
NG_number(data, whiten = c("sqrtprec", "eigenvec"))
}
\arguments{
\item{data}{original matrix with n x p.}

\item{whiten}{'sqrtprec' or 'eigenvec'}
}
\value{
the number of non-Gaussian components in the data.
}
\description{
find the number of non-Gaussian components in the data.
}
\examples{
\dontrun{
library(singR)
data("exampledata")
data=exampledata
NG_number(data$dX)
}
}
