% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_functions.r
\name{signchange}
\alias{signchange}
\title{Sign change for S matrix to image}
\usage{
signchange(S, M = NULL)
}
\arguments{
\item{S}{S, r x px.}

\item{M}{Mx, n x r.}
}
\value{
alist of positive S and positive Mx.
}
\description{
Sign change for S matrix to image
}
