% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic functions.r
\name{signchange}
\alias{signchange}
\title{Sign change for S matrix to image}
\usage{
signchange(S, M = NULL)
}
\arguments{
\item{S}{t(S) px x r.}

\item{M}{another matrix.}
}
\value{
newS the original S * signskew, which is mean(S.column^3)
}
\description{
Sign change for S matrix to image
}
