% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_functions.r
\name{whitener}
\alias{whitener}
\title{Whitening Function}
\usage{
whitener(X, n.comp = ncol(X), center.row = FALSE)
}
\arguments{
\item{X}{dataset p x n.}

\item{n.comp}{the number of components}

\item{center.row}{whether center the row of data}
}
\value{
a whitener matrix
}
\description{
Whitening Function
}
